requirements.txt
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
pdfminer.six==20221105
pycparser==2.21
pycryptodome==3.17
PyPDF2==3.0.1
